refactor(log_view): change code to work with predefined types
Since we have a types file, it's best to use them to stay up to
date with the backend's way of sending back information.

A few things were noted in the code as comments, though. Such as
having to handle optional information, such as User.email and
User.phone_number. I added a lot of helper functions too, though
idk if they should be kept in this same file.