chore(backend): create Medication.refill(), .total_prescribed
Medications now include a total_prescribe attribute to use for the
circle UI element. That element can only be changed with .refill().

A note about .refill() though: the current supply of a medication
will be added to the .refill() "supply" param, so it is possible
that Medication.supply is greater than Medication.total_prescribed.