chore: testing Bun
Bun is actually much easier to transition to than Deno. Everything
seems to work out of the box. Though, I'm not sure if it's worth
transitoning to anything at the moment? We're saving time, sure,
but I think it's best to transition stuff when building a brand
new project.

It's almost as if I could've figured this out without coding this.

:^)