fix(medications): Medication.brand made optional
I want to do some more research when it comes to how medications
are usually stored on databases. I'm not sure brand is ever
included, at least in a mandatory fashion.

This also fixes an issue where the Submit button actually
disables itself instead of just styling itself as disabled.