refactor(core): split all core-related functionality
Having "core" folders in the program led to confusion when it
came to development, and even caused functionality between
tools to be wrapped around this core folder. This commit
makes drastic changes to Iris' functionality.

This commit does the following:

- Split "core" folder into settings, onboarding, hub, etc.
- "User" struct/OOP patterns are deleted, a user is no longer
  needed to fetch medications, isolating the tool
- Users will be kept track of with the Contacts tool
  - The first contact to be created will be the patient, which
    will not be allowed to be deleted
- Makes the hub its own tool in the frontend
- Prepares folders for a new onboarding experience
- Relay section in Settings temporarily removed
- Removed various config options, moved to settings
- Various helper functions moved to root of frontend/backend

This commit will likely be destructive, and bugs should be
expected. This will also undoubtably cause merge conflicts,
so developers should be on the lookout. 🫡